+++ OrderManagementService/src/main/java/com/company/ordermanagementservice/model/Order.java
@@ -16,7 +16,7 @@ public class Order {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private BigInteger id;
 
     @Column(name = "user_id")
     private BigInteger userId;
