+++ OrderManagementService/src/main/java/com/company/ordermanagementservice/repository/OrderItemRepository.java
@@ -3,5 +3,7 @@ package com.company.ordermanagementservice.repository;
 import com.company.ordermanagementservice.model.OrderItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderItemRepository extends JpaRepository<OrderItem, Long> {
+import java.math.BigInteger;
+
+public interface OrderItemRepository extends JpaRepository<OrderItem, BigInteger> {
 }
