+++ OrderManagementService/src/main/java/com/company/ordermanagementservice/repository/OrderRepository.java
@@ -3,5 +3,7 @@ package com.company.ordermanagementservice.repository;
 import com.company.ordermanagementservice.model.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderRepository extends JpaRepository<Order, Long> {
+import java.math.BigInteger;
+
+public interface OrderRepository extends JpaRepository<Order, BigInteger> {
 }
